pkp/pkp-lib#13156 Reorganize how the useReviewDetails apis are structured
